# docs/openapi-agent-skills.yaml
openapi: "3.1.0"
info:
  title: DSG SkillGate — Agent Skills API
  version: "1.0.0"
  description: |
    Open-source GitHub skill discovery, inspection, verification, locking, and gated execution.

    Pipeline: Search → Inspect → Build Draft → Verify Gate → Lock Registry → Run via Gate

    **Safety invariants (enforced at runtime):**
    - Read-only GitHub metadata access — no code is cloned, installed, or executed
    - `blocked` skills cannot be registered without `forceRegister: true`
    - `needs_approval` / `needs_review` skills are denied at the run gate
    - All execute actions pass through `prepareGovernedToolRequest()` for audit
  contact:
    name: DSG ONE
    url: https://dsg.pics
servers:
  - url: /api/agent-skills
    description: Next.js App Router (same origin)
security:
  - BearerAuth: []
  - WorkspaceHeader: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Supabase user access token
    WorkspaceHeader:
      type: apiKey
      in: header
      name: x-dsg-workspace-id
  schemas:
    SkillRiskLevel:
      type: string
      enum: [low, medium, high, critical]
...
    SkillPermissions:
      type: object
      properties:
        network:
          type: boolean
        filesystem:
          type: string
          enum: [none, read, write]
        secrets:
          type: boolean
        codeExecution:
          type: boolean
        externalWrite:
          type: boolean
      required: [network, filesystem, secrets, codeExecution, externalWrite]
...
    SkillDraft:
      type: object
      properties:
        id:
          type: string
          description: Slugified identifier from owner/repo
          example: anthropics-claude-code
        name:
          type: string
        sourceType:
          type: string
          enum: [github, npm, local]
        sourceUrl:
          type: string
          format: uri
        sourceOwner:
          type: string
        sourceRepo:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [draft]
        riskLevel:
          $ref: "#/components/schemas/SkillRiskLevel"
        permissions:
          $ref: "#/components/schemas/SkillPermissions"
        draftedAt:
          type: string
          format: date-time
    SkillVerificationResult:
      type: object
      properties:
        status:
          type: string
          enum: [verified, needs_review, needs_approval, blocked]
        reasons:
          type: array
          items:
            type: string
          description: |
            Flags explaining the status. Possible values:
            MISSING_README, MISSING_LICENSE, NO_TESTS_FOUND,
            HARDCODED_SECRET_DETECTED, EXTERNAL_WRITE_CAPABILITY,
            CODE_EXECUTION_CAPABILITY
        checks:
          type: object
          properties:
            hasReadme:
              type: boolean
            hasLicense:
              type: boolean
            hasTests:
              type: boolean
            noHardcodedSecrets:
              type: boolean
            noExternalWrite:
              type: boolean
            noCodeExecution:
              type: boolean
            riskLevel:
              $ref: "#/components/schemas/SkillRiskLevel"
...
    ErrorResponse:
      type: object
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          properties:
            code:
              type: string
paths:
...
  /build:
    post:
      summary: Build a skill draft with risk assessment
      operationId: buildSkillDraft
      security:
        - BearerAuth: []
          WorkspaceHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner, repo]
              properties:
                owner:
                  type: string
                repo:
                  type: string
      responses:
        "200":
          description: Skill draft with risk and permissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      draft:
                        $ref: "#/components/schemas/SkillDraft"
                      inspectionSummary:
                        type: object
                      nextStep:
                        type: string
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /verify:
    post:
      summary: Run verification gate on a skill draft
      description: |
        Accepts either a pre-built draft (from /build) or owner+repo to build+verify inline.
        Returns one of: verified | needs_review | needs_approval | blocked
      operationId: verifySkill
      security:
        - BearerAuth: []
          WorkspaceHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - type: object
                  required: [draft]
                  properties:
                    draft:
                      $ref: "#/components/schemas/SkillDraft"
                - type: object
                  required: [owner, repo]
                  properties:
                    owner:
                      type: string
                    repo:
                      type: string
      responses:
        "200":
          description: Verification decision
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                    enum: [true]
                  data:
                    type: object
                    properties:
                      skillId:
                        type: string
                      verification:
                        $ref: "#/components/schemas/SkillVerificationResult"
                      draft:
                        $ref: "#/components/schemas/SkillDraft"
                      nextStep:
                        type: string
        "403":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...